+++ src/csp/domain/intdom.rs
@@ -1,5 +1,4 @@
 /**************************************
 - Author: Clement Poncelet
-- Desc: Variable, a string label associated with a domain<T> (set of possible values)
-        ExVar because it is for now implemented using extensional domains
+- Desc: Interval Domain, representing a set of values with [min, max]
 ***************************************/