+++ src/csp/domain/mod.rs
@@ -1,2 +1,2 @@
 pub mod traits;
-pub mod extdom;
+pub mod setdom;