+++ src/csp/variable/extvar.rs
@@ -9,7 +9,7 @@
             Factories
 ***************************************/
 
-pub fn generate_variables<T:OrdT>(base_name: &str, n:usize, dom : &ExDom<T>) -> HashMap<String, Rc<ExVar<T>>> {
+pub fn generate_variables<T:OrdT>(base_name: &str, n:usize, dom : &SetDom<T>) -> HashMap<String, Rc<ExVar<T>>> {
     let mut vmap = HashMap::new();
     for i in 1..n+1 {
         vmap.insert(
@@ -23,35 +23,45 @@ pub fn generate_variables<T:OrdT>(base_name: &str, n:usize, dom : &ExDom<T>) -> 
 /**************************************
             Variables
 ***************************************/
+use std::cell::RefCell;
 use std::collections::HashMap;
 use std::fmt;
 use std::rc::Rc;
-use crate::csp::domain::extdom::ExDom;
+use crate::csp::domain::setdom::SetDom;
 use crate::csp::domain::traits::{Domain, OrdT};
 use crate::var;
 
 pub struct ExVar<T:OrdT> {
     label: String,
-    dom: ExDom<T>
+    dom:  Rc<RefCell<SetDom<T>>>
 }
 
 impl<T:OrdT> ExVar<T> {
-    pub fn new (label: String, dom: ExDom<T>) -> ExVar<T> {
-        ExVar { label, dom }
+    pub fn new (label: String, dom: SetDom<T>) -> ExVar<T> {
+        let ref_dom = Rc::new(RefCell::new(dom));
+        Self {
+            label,
+            dom: ref_dom
+        }
     }
     pub fn value(&self) -> Option<T> {
-        if self.dom.size() > 1  {None}
-        else {Some(self.dom.get().get(0).unwrap().clone())}
+        if self.dom().size() > 1  {None}
+        else {Some(self.dom().active_values().get(0).unwrap().clone())}
     }
-    pub fn valid_values(&self) -> &Vec<T> { self.dom.get() }
-    pub fn valid_size(&self) -> usize { self.dom.size() }
+    pub fn valid_values(&self) -> Vec<T> { self.dom().active_values() }
+    pub fn valid_size(&self) -> usize { self.dom().size() }
     pub fn label(&self) -> &String {&self.label}
-    pub fn dom(&self) -> &ExDom<T> {&self.dom}
+    pub fn dom(&self) -> std::cell::Ref<'_, SetDom<T>> {
+        self.dom.borrow()
+    }
+    pub fn dom_mut(&self) -> std::cell::RefMut<'_, SetDom<T>> {
+        self.dom.borrow_mut()
+    }
 }
 
 impl<T:OrdT> fmt::Display for ExVar<T> {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{} : {}", self.label, self.dom)
+        write!(f, "{} : {}", self.label, self.dom())
     }
 }
 
@@ -61,11 +71,13 @@ impl<T:OrdT> fmt::Display for ExVar<T> {
 
 #[cfg(test)]
 mod tests {
-    use crate::csp::variable::extvar::ExDom;
+    use std::ops::Deref;
+    use crate::csp::domain::traits::Domain;
+    use crate::csp::variable::extvar::SetDom;
     use crate::csp::variable::extvar::ExVar;
 
-    fn int_dom() -> ExDom<i32> {
-        ExDom::new(vec![1, 2, 3])
+    fn int_dom() -> SetDom<i32> {
+        SetDom::new(vec![1, 2, 3])
     }
 
     #[test]
@@ -77,10 +89,10 @@ mod tests {
     #[test]
     fn exvar_domain_access() {
         let dom = int_dom();
-        let x = ExVar::new("x".to_string(), dom.clone());
+        let x = ExVar::new("x".to_string(), dom.snapshot());
 
-        assert_eq!(x.dom(), &dom);
-        assert_eq!(x.valid_values(), &vec![1, 2, 3]);
+        assert_eq!(x.dom().deref(), &dom);
+        assert_eq!(x.valid_values(), vec![1, 2, 3]);
         assert_eq!(x.valid_size(), 3);
     }
 
@@ -90,4 +102,44 @@ mod tests {
         assert_eq!(x.value(), None);
     }
 
+    #[test]
+    fn exvar_dom_read() {
+        let dom = SetDom::new(vec![1, 2, 3]);
+        let x = ExVar::new("x".into(), dom);
+
+        let values = x.dom().active_values();
+        assert_eq!(values, vec![1, 2, 3]);
+    }
+
+    #[test]
+    fn exvar_dom_mut_remove_value() {
+        let dom = SetDom::new(vec![1, 2, 3]);
+        let x = ExVar::new("x".into(), dom);
+
+        {
+            let mut d = x.dom_mut();
+            d.remove_value(&2, 0);
+        }
+
+        let values = x.dom().active_values();
+        assert_eq!(values, vec![1, 3]);
+    }
+
+    #[test]
+    fn exvar_snapshot_distinct_domain() {
+        let dom = SetDom::new(vec![1, 2, 3]);
+        let x = ExVar::new("x".into(), dom.snapshot());
+        let y = ExVar::new("y".into(), dom);
+
+        {
+            let mut dx = x.dom_mut();
+            dx.remove_value(&1, 0);
+        }
+
+        let vx = x.dom().active_values();
+        let vy = y.dom().active_values();
+
+        assert_eq!(vx, vec![2, 3]);
+        assert_eq!(vy, vec![1, 2, 3]);
+    }
 }